[feat] #110 기기대응을 위한 폰트, 패딩값 수정 함수 파일 추가, 온보딩 파일에서 익스텐션 삭제
