+++ Semo/Views/OnBoarding/OnBoardingView.swift
@@ -60,25 +60,6 @@ struct OnBoardingView: View {
     
 }
 
-
-extension UIScreen {
-    
-    static let screenWidth = UIScreen.main.bounds.size.width
-    static let screenHeight = UIScreen.main.bounds.size.height
-    static let screenSize = UIScreen.main.bounds.size
-
-    static func getWidth(_ width: CGFloat) -> CGFloat {
-        screenWidth / 390 * width
-    }
-
-    static func getHeight(_ height: CGFloat) -> CGFloat {
-        screenHeight / 844 * height
-    }
-}
-
-
-
-
 struct OnBoardingView_Previews: PreviewProvider {
     static var previews: some View {
         OnBoardingView()
